Bug 1669181 - Part 50: Split JSAtomState from "vm/Runtime.h". r=jandem
Split `JSAtomState` from "vm/Runtime.h", so that "jit/CompileInfo.h" doesn't
need to pull in everything from "vm/Runtime.h".

Drive-by change:
- Remove an unnecessary include from "vm/CommonPropertyNames.h".

Differential Revision: https://phabricator.services.mozilla.com/D92743
